Article Additions
"Article Guide" document to help fill out each section, as per the assignment spec.

"References" document to keep track of references along with citation number until we can index it ourselves.

Small changes to the paper to make it less informal. Still more needed on this.

Small additions to Introduction to argue "novel contribution" and high level overview of the game in the "System Design" section.