added Timer, Paper - Implementation, References; changed addPlayer
Regular players can now detect if a leader has timed out and begin election (currently commented out as election still unimplemented). addPlayer() now returns a GameInt object not a boolean and a Player is correctly set as a role if a Leader is present. This was needed as the solution added during the last meeting did not work. Some content was added to the Implementation section of the paper. Two references were added, one is from Youtube, I hope it works with the current syntax.